UPSTREAM: <carry>: ignore drift on unstructured ignition userdata token field
We need to specially hash EC2NodeClass custom userData.
This is because the karpenter-operator passes a rotating token to the EC2NodeClass userData field as part of the payload.
This field unintentionally causes the hash to change and trigger drift.

This happens by parsing a special header we include in ignition-server requests, and only returning that header TargetConfigVersionHash's value.
The hash is unique and will act as a trigger for Drift rollout, similar to it's usage in HyperShift and the hyperv1.NodePool API.
https://github.com/openshift/hypershift/blob/07b5bf9a97d23d6c7a01164a385e5b9d6c513794/hypershift-operator/controllers/nodepool/config.go#L120

If the config was not a valid ignition config, then we bypass this special handling entirely.

Signed-off-by: Max Cao <macao@redhat.com>