UPSTREAM: <carry>: Add aws-karpenter-provider-aws to payload
Eventually we need to do this over in hypershift, but this label tells
oc to include us as an operator in OpenShift. Without it, our image will
be ignored and not included. Yes, this was in here before, but a bad
squash took it out of the 1.8.2 rebase unintentionally, so this just
puts it back until we can make hypershift register its consumption of
this image properly.

Signed-off-by: John Kyros <jkyros@redhat.com>